Fix:core:Fix some warnings
git-svn-id: http://svn.code.sf.net/p/navit/code/trunk/navit@5032 ffa7fe5e-494d-0410-b361-a75ebd5db220

+++ navit/maptool/osm_protobufdb.c
@@ -421,7 +421,7 @@ osm_protobufdb_string(struct osm_protobufdb_context *ctx, char *str)
 	gpointer value;
 	assert(ctx->string_hash != NULL);
 	if (g_hash_table_lookup_extended(ctx->string_hash, str, NULL, &value)) {
-		return (int)value;	
+		return (long)value;
 	}
 	if (!st->n_s) {
 		st->n_s++;
